fix storage eror

+++ client/src/helpers/SessionStorageHelper.ts
@@ -1,24 +1,34 @@
 export const storeString = (key: keys, data: string) => {
-  sessionStorage.setItem(key, data);
+  if(typeof(Storage) !== "undefined"){
+    sessionStorage.setItem(key, data);
+  }
 };
 
 export const storeObject = (key: keys, data: any) => {
-  sessionStorage.setItem(key, JSON.stringify(data));
+  if(typeof(Storage) !== "undefined"){
+    sessionStorage.setItem(key, JSON.stringify(data));
+  }
 };
 
 export const getObject = (key: keys): any => {
-  return JSON.parse(<string>sessionStorage.getItem(key));
+  if(typeof(Storage) !== "undefined"){
+    return JSON.parse(<string>sessionStorage.getItem(key));
+  }
 };
 
-export const getString = (key: keys): string => {
-  return <string>sessionStorage.getItem(key);
+export const getString = (key: keys): any => {
+  if(typeof(Storage) !== "undefined"){
+    return <string>sessionStorage.getItem(key);
+  }
 };
 
 export const removeItems = (key?: keys) => {
-  if (key != null) {
-    sessionStorage.removeItem(key);
-  } else {
-    sessionStorage.clear();
+  if(typeof(Storage) !== "undefined"){
+    if (key != null) {
+      sessionStorage.removeItem(key);
+    } else {
+      sessionStorage.clear();
+    }
   }
 };
